requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
garth>=0.4.0
fitdecode>=0.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
cryptography>=42.0.0
